refactor(mbk): drop local copies of auth_event_service and admin_user_service_factory
Removes two parity-drift sources flagged in the 2026-05-08 monorepo
refactor audit:

- app/services/system/auth_event_service.py — single-line passthrough
  (`from platform_shared.services.auth_event_service import log_auth_event`).
  Five consumers updated to import from platform_shared directly:
  api/totp.py, core/auth.py, core/rate_limit.py,
  services/integrations/integration_service.py,
  services/user/account_service.py.

- app/services/system/admin_user_service_factory.py — DI wiring stub
  that instantiated AdminUserService with MBK's User model and session
  factories. The singleton (shared_admin_user_service) moved to
  admin_service.py where it is primarily consumed. Three consumers
  updated: app/main.py, services/system/admin_service.py (now owns the
  instantiation), tests/test_admin_routes.py.

app/services/user/totp_service.py is intentionally kept — it contains
non-trivial MBK-specific logic (MyBookkeeper TOTP issuer string,
settings.encryption_key binding, and all DB-coupled coordinators:
setup_totp, confirm_totp, disable_totp, validate_totp_for_login,
is_totp_required). The shared module's own docstring explicitly states
"the DB-coupled half stayed in MBK as a thin orchestration layer."
Deleting it is a separate decision requiring a different consolidation
shape. See report for details.

All 100 targeted tests pass (totp + auth_event + admin_user domains);
58/58 auth-adjacent smoke tests pass.

Resolves "MBK has local copies of code already in platform_shared"
(partial — auth_event and admin_user_service_factory items) from the
2026-05-08 monorepo refactor audit.
