+++ apps/mybookkeeper/backend/app/api/totp.py
@@ -26,7 +26,7 @@ from app.schemas.user.totp import (
     TotpVerifyRequest,
     TotpVerifyResponse,
 )
-from app.services.system.auth_event_service import log_auth_event
+from platform_shared.services.auth_event_service import log_auth_event
 from app.services.user import totp_service
 
 router = APIRouter(prefix="/auth/totp", tags=["auth"])
